Solves #26
