--- client/src/pages/recipes/index.jsx
@@ -1,3 +1,0 @@
-export function RecipesPage() {
-  return <>recipes</>
-}
